compat-wireless: do not add const before SIMPLE_DEV_PM_OPS
For kernel < 2.6.32 .driver.pm in struct pci_driver is not const, but
for later kernel versions it is const. This const should not be added
manually, but by SIMPLE_DEV_PM_OPS(), so remove it from this patch.

In addition compat_pci_suspend and compat_pci_resume are now in front
of SIMPLE_DEV_PM_OPS to make make it possible to use a function created
by them in SIMPLE_DEV_PM_OPS(), which is needed for kernel 2.6.29.

This is partly based on work done by Johannes Berg.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>
Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>