compat-drivers: Add ability to select drivers to fetch
Add a last command line parameter to define which subsystems will
be fetched into the tree.

Currently the script supports network and/or drm keywords as the
last argument.

network is currently enabled by default to retain compatibility with
the previous compat-wireless tree.

Move around code and use the previously added 2 helper functions
to copy files/folders conditionally, according to the values of
ENABLE_NETWORK (and ENABLE_DRM which will follow).

Signed-off-by: Ozan Çağlayan <ozancag@gmail.com>
Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>