compat-drivers: add kup support with git-paranoia support
kup [0] is the kernel.org uploader. This integrates support for
usage of kup with the gen-release.sh script. When kup is used
we also use git-paranoia which will bail out if either we have
content on a git tree that lets us produce content that is not
GPG signed, that has not yet been commited, or integrated into
a git tree. To use kup just pass on the -u argument to
gen-release.sh.

For generic releases, that do not use kup (without -u) we still
run git-paranoia but in interactive mode, allowing releases to
be made that are not targeted for kernel.org with random junk
or not GPG signed junk.

[0] https://korg.wiki.kernel.org/index.php/Userdoc:using_kup

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>