compat-drivers: enable DRM drivers compilation
Ozan Çağlayan had done most of the work to suck in all
the DRM drivers but left out the enablement of compiling
the drm drivers. The DRM drivers were also not load
tested. This patch enables to test compilation by
renaming the DRM Kconfig options with a _COMPAT prefix
and enables compilation by default for the i915 drm
driver which I can test against at run time on my
laptop running on v3.2.

Users and developer can help enable older kernels
down to 3.0 (as this was the oldest supported target
kernel for DRM drivers) or other DRM drivers as they
are run time tested.

The admin-update.sh script is modified to now always
run by including the DRM code and patching it accordingly.
If you happen to only want networking code you can do so
by supplying the network argument to the script as follows:

./scripts/adming-update.sh network

This will disable taking in DRM code.

It should be noted that although only the i915 drm driver
is being enabled all other drivers except the following have
been test compiled to work down to v3.0.

-obj-$(CONFIG_COMPAT_DRM_TDFX)  += tdfx/
-obj-$(CONFIG_COMPAT_DRM_R128)  += r128/
-obj-$(CONFIG_COMPAT_DRM_MGA)   += mga/
-obj-$(CONFIG_COMPAT_DRM_SIS)   += sis/
-obj-$(CONFIG_COMPAT_DRM_SAVAGE)+= savage/
-obj-$(CONFIG_COMPAT_DRM_EXYNOS) +=exynos/
-obj-$(CONFIG_COMPAT_DRM_UDL) += udl/

Apart from the above drivers we should be able to enable the
other drivers as reports come in for run time testing on them.
As it is right now only the i915 goes run time tested -- by me.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>