compat-drivers: prevent mixing patches from different directories in refresh
The refresh script is able to take an already existing quilt run and
just refresh the further needed patches. This is needed to make it
possible to manually fix a patch which is not applying any more and
then automatically integrate it into the patches directory again. With
the integration of drm it could happen that quilt is running on two
different source directories like patches/network and patches/drm in one
call of admin-update.sh. If some of the drm patches did not apply and
needed a manual fix, quilt moved the patch in the second run into the
patches/network directory, because it was taken in the network refresh
run. This patch does not run quilt if there is still an unfinished
quilt job for a different patch directory.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>
Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>