compat-drivers: run refresh for all patches
In the past refresh with quilt was just run for the last patch
directory being applied and not for all, because our script had
problems when patchRefresh() was called more then once in a single
admin-update.sh run. This was fixed in the last patch and now it is
safe to run it on all patches.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>
Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>