compat-drivers: move all patch types under patches/
The number of patch directories for the different life
cycle a patch can be on its way upstream had reached 4:

  * pending-stable
  * linux-next-pending
  * linux-next-cherry-picks
  * crap

Apart from this we also had the backport related patches.
In the future I see potential for other directories for
different categories of patches. Consider how we are
breaking down the backport patches into specific documented
collateral evolutions and of those type of collateral
evolutions there may be different types of collateral
evolutions which we could likely treat / document differently.
Another type of patch that I forsee us having to deal with
are patches for drivers that are not upstrea -- like crap,
but for drivers in a unified driver tree that is *designed*
from scratch to be targetting upstream Linux inclusion using
the same coccinelle technology I forsee us using for backporting
collateral evolutions.

Since the number of patch directories can get out of hand
at the top level lets just stuff then all under patches/
and deal with patches there. This also places the backport
patches into a directory called collateral-evolutions/ to
help promote and educate on the nature of the changes
that we have to deal with when backporting.

Cc: Julia Lawall <julia.lawall@lip6.fr>
Cc: Adrian Chadd <adrian@freebsd.org>
Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>