compat-drivers: add support for unified drivers
Companies typically write Linux device drivers either from
scratch of alternatively based on some internal unified
source code tree. Typically the unified driver architecture
is a mess and porting a driver to linux is always a painful
task to the extent at times a full rewrite is required and
at times preferred.

In order to try to help with these type of unified driver
strategies one possibility that has been proposed is to
make the unified driver strategy an open source project and
to not make many compromises on the style and requirements
for Linux. The alx Linux driver [1] is not yet upstream and
is undergoing changes to try to accomodate it to a format
and use new technologies to help address unification attempts
in both an open manner and also lead by the community.

compat-drivers will support these type of driver efforts
only if they adhere to the principles of never compromising
Linux upstream requirements and are targetting getting
upstream into the Linux kernel. Once these drivers are upstream,
their respective git trees will be removed and we'll simply
cherry pick the driver as we do from upstream Linux. If a
compat-drivers releases contains drivers that are from a
unified git repository the release will be postfixed with
a "-u".

[1] http://www.linuxfoundation.org/collaborate/workgroups/networking/alx

Cc: Julia Lawall <julia.lawall@lip6.fr>
Cc: Adrian Chadd <adrian@freebsd.org>
Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>