compat-drivers: add the MDIO module
The alx Ethernet driver depends on the MDIO module.
This module has UAPI and kernel headers. The headers are
carried over intact to help *both* backport it for older
drivers and also provide the newer driver to other kernels
that may have it enabled after it was introduced in 2.6.31.

The way support for MDIO is enabled is by carrying over both
the UAPI and Linux kernel header files along with the module
with in compat-drivers. The exported symbols of the module
however are renamed with a compat_* prefix within compat
in include/linux/compat-2.6.31.h. This enables Linux
distributions shipping a driver that depends on older
versions of the MDIO to keep using it while allowing the
supplied drivers in compat-drivers to become users of
the newer module it likely will require a tighter
dependency on.

An alternative to using different exported symbol names
is for compat-drivers to work under a separate namespace
and while I hear some folks already implemented that support
we never got those patches sent to us.

C'est la vie.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>