compat-drivers: fix collateral-evolutions/network/53-pr_fmt.patch
The brcmfmac driver was updated recently to only define
pr_fmt() once and it also had the undef for us. Technically
Joe should not have had to add the undef as I previously
went on a crusade to verify that this was not needed at the
driver level so an upstream patch is welcomed to remove that
undef upstream and then to add it for backporting purposes.

commit 79d7c4e8da4ac3cfa98e8e622a17baaeb3c9d29d
Author: Arend van Spriel <arend@broadcom.com>
Date:   Wed Jan 2 21:20:10 2013 +0100

    brcmfmac: define pr_fmt in one place

    Several source files (but not all) define the pr_fmt() macro in exactly
    the same way. Instead this commit defines it in a header file so driver
    logging is consistent.

    Cc: Joe Perches <joe@perches.com>
    Reviewed-by: Hante Meuleman <meuleman@broadcom.com>
    Reviewed-by: Pieter-Paul Giesberts <pieterpg@broadcom.com>
    Signed-off-by: Arend van Spriel <arend@broadcom.com>
    Signed-off-by: John W. Linville <linville@tuxdriver.com>

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>