compat-drivers: fix driver-select for atheros drivers
Upstream commit 3bd8075f0152188a315fb092510d66b38de014ea

The addition of the new wil6210 driver added a new ATH_CARDS
option to be used to describe Atheros cards and now uses ATH_COMMON
for the common module ath. This requires fixing the driver-select
script so that the drivers are enabled properly.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>