compat-drivers: rename disable-dump-adjust-on-old-kernels.patch
Rename this from:

70-disable-dump-adjust-on-old-kernels.patch

to:

0006-disable-dump-adjust-on-old-kernels.patch

4 digit patches are patches which have undergone review for
breaking it down as much as possible to help with seeing if
some of these can later be expressed as SmPL grammer for usage
with Coccinelle.

At this point SmPL grammer is possible for 0001-netdev_ops.patch
but that is because we have all the other 2 digit patches yet to
break down further.

This patch submitted by Johannes is already well broken down
so rename it to annotate it has been reviewed and is as atomic
as can be.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>