compat-drivers: add initial kconfig development
This adds intial development kconfig support (make menuconfig).
Synching in kconfig from linux-next next-20130326.
Support is still under development so to enable it
if you want to work on it, just run manually:

  make -f scripts/kconfig/Makefile menuconfig

compat already provides its own set of scripts that
computes the kernels you need and spits it out to .config
through the call:

  ./compat/scripts/gen-compat-config.sh > $(COMPAT_CONFIG)

Perhaps something along these lines will do it:

	--- a/Makefile
	+++ b/Makefile
	@@ -29,6 +29,7 @@ DESTDIR?=
	 ifneq ($(KERNELRELEASE),)

	 -include $(COMPAT_CONFIG)
	+-include $(BACKPORT_CONFIG)
	 include $(COMPAT_CONFIG_CW)

	 NOSTDINC_FLAGS := \
	@@ -86,7 +87,8 @@ export CFLAGS += \
	 # These exported as they are used by the scripts
	 # to check config and compat autoconf
	 export COMPAT_CONFIG_CW=$(PWD)/config.mk
	-export COMPAT_CONFIG=$(PWD)/.config
	+export COMPAT_CONFIG=$(PWD)/compat/.config
	+export BACKPORT_CONFIG=$(PWD)/.config

But I'm done for the day.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>